// src/models/JobAPI.ts
export default interface JobContentAPI {
  intitule_job: string
  description: string
  image: string
  offre_description: string
  offre_list_point1: string
  offre_list_point2: string
  offre_list_point3: string
  offre_list_point4: string
  offre_list_point5: string
  offre_list_point6: string
  offre_list_point7: string
  offre_list_point8: string
  offre_list_point9: string
  offre_list_point10: string
  offre_profile_description: string
  offre_salary: string
  offre_list_profile1: string
  offre_list_profile2: string
  offre_list_profile3: string
  offre_list_profile4: string
  offre_list_profile5: string
  offre_list_profile6: string
  offre_last_paragraph: string
  slug: string
  contact_text: string
}